requirements.txt
Django==4.1.6
django_filter==22.1
djangorestframework==3.14.0
drf_yasg==1.21.4
black
